Correct cargo hack errors
