+++ src/ecies/generic_ecies_aes128gcm.rs
@@ -245,7 +245,7 @@ impl<
     }
 }
 
-#[cfg(all(test, feature = "nist_curves"))]
+#[cfg(all(test, feature = "nist_curves", feature = "curve_25519"))]
 mod tests {
     use aead::Payload;
